Re-namespace multiplayer requests/responses
