+++ osu.Game/Online/Multiplayer/PartRoomRequest.cs
@@ -3,15 +3,15 @@
 
 using System.Net.Http;
 using osu.Framework.IO.Network;
-using osu.Game.Online.Multiplayer;
+using osu.Game.Online.API;
 
-namespace osu.Game.Online.API.Requests
+namespace osu.Game.Online.Multiplayer
 {
-    public class JoinRoomRequest : APIRequest
+    public class PartRoomRequest : APIRequest
     {
         private readonly Room room;
 
-        public JoinRoomRequest(Room room)
+        public PartRoomRequest(Room room)
         {
             this.room = room;
         }
@@ -19,7 +19,7 @@ namespace osu.Game.Online.API.Requests
         protected override WebRequest CreateWebRequest()
         {
             var req = base.CreateWebRequest();
-            req.Method = HttpMethod.Put;
+            req.Method = HttpMethod.Delete;
             return req;
         }
